requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
openai-whisper
gTTS
python-multipart
pydub